# refactor: eliminate `notify` recursion