@johnsoncodehk 2025-08-01
Since the notify function isn't a performance hotspot, I think keeping the Surface API code simple and easier for people to understand will have a greater benefit.

Thanks for the PR!